revert(streamflow): Remove streamflow instruction from proposals page
