+++ hooks/useGovernanceAssets.ts
@@ -144,16 +144,6 @@ export default function useGovernanceAssets() {
       isVisible: canUseTokenTransferInstruction,
     },
     {
-      id: Instructions.CreateStream,
-      name: 'Streamflow: Create Vesting Contract',
-      isVisible: canUseAnyInstruction,
-    },
-    {
-      id: Instructions.CancelStream,
-      name: 'Streamflow: Cancel Vesting Contract',
-      isVisible: canUseAnyInstruction,
-    },
-    {
       id: Instructions.Grant,
       name: 'Grant',
       isVisible:
