+++ pages/dao/[symbol]/proposal/new.tsx
@@ -66,8 +66,6 @@ import FriktionClaimPendingWithdraw from './components/instructions/Friktion/Fri
 import MakeChangePerpMarket from './components/instructions/Mango/MakeChangePerpMarket'
 import MakeAddOracle from './components/instructions/Mango/MakeAddOracle'
 import MakeAddSpotMarket from './components/instructions/Mango/MakeAddSpotMarket'
-import CreateStream from './components/instructions/Streamflow/CreateStream'
-import CancelStream from './components/instructions/Streamflow/CancelStream'
 import StakeValidator from './components/instructions/Validators/StakeValidator'
 import DeactivateValidatorStake from './components/instructions/Validators/DeactivateStake'
 import WithdrawValidatorStake from './components/instructions/Validators/WithdrawStake'
@@ -432,8 +430,6 @@ const New = () => {
       [Instructions.MangoV4Serum3RegisterMarket]: Serum3RegisterMarket,
       [Instructions.MangoV4PerpCreate]: PerpCreate,
       [Instructions.MangoV4TokenRegisterTrustless]: TokenRegisterTrustless,
-      [Instructions.CreateStream]: CreateStream,
-      [Instructions.CancelStream]: CancelStream,
       [Instructions.Grant]: Grant,
       [Instructions.Clawback]: Clawback,
       [Instructions.CreateAssociatedTokenAccount]: CreateAssociatedTokenAccount,
